weird error for a particular image but rest done

+++ image_labeling/manufacture_data.py
@@ -10,12 +10,14 @@ DEBUG = False
 # Initial values
 empty_imgs_dir = 'empty_imgs'
 flu_odlcs_dir = 'flu_odlcs'
+out_dir = 'labeled_dataset'
 cwd = os.getcwd()
 empty_imgs = os.listdir(os.path.join(cwd, empty_imgs_dir))
 flu_odlcs = os.listdir(os.path.join(cwd, flu_odlcs_dir))
 
 # Max number
 num_obj_in_img = 10
+odlc_scale = 8
 
 # Modified imutils rotate_bound function
 # Simply changing how the warpaffine function is called
@@ -48,7 +50,7 @@ def rotate_bound(image, angle):
             cv.INTER_LINEAR, 
             borderMode=cv.BORDER_CONSTANT, 
             borderValue=(255,255,255)
-            )
+            ), M, (nW, nH)
 
 
 # Initialise the info about each odlc
@@ -75,37 +77,57 @@ for flu_shape in flu_odlcs:
     odlc_arr = cv.imread(odlc_path)
     cv.imshow(f"{meta['file_name']}", odlc_arr)
     cv.waitKey(0)
+    meta['icdar_text'] = {}
     while not qu.empty():
-        meta['char_top_left'] = qu.get(False)
-        meta['char_bottom_right'] = qu.get(False)
+        meta['icdar_text']['xy1'] = qu.get(False)
+        if not qu.empty():
+            meta['icdar_text']['xy2'] = qu.get(False)
+        if not qu.empty():
+            meta['icdar_text']['xy3'] = qu.get(False)
+        if not qu.empty():
+            meta['icdar_text']['xy4'] = qu.get(False)
     cv.destroyAllWindows()
     print(meta)
     li.append(meta)
 
 
 flu_odlcs = li
-
-for img in empty_imgs:
+num_images = 0
+for img in empty_imgs[850:860]:
+    num_images += 1
+    print(num_images)
+    print('going through', img)
     img_path = os.path.join(cwd, empty_imgs_dir, img)
     img_arr = cv.imread(img_path)
 
     added_odlcs = random.sample(flu_odlcs, random.randint(0,num_obj_in_img))
-    
+    print('Adding odls:', added_odlcs, 'to', img)
     for odlc in added_odlcs:
         odlc_path = os.path.join(cwd, flu_odlcs_dir, odlc['file_name'])
+        curr_odlc_text_loc = odlc['icdar_text'].copy()
 
         odlc_arr = cv.imread(odlc_path)
         if DEBUG: print(odlc_arr, odlc_arr.shape)
-        scale = 8
+        scale = odlc_scale
+        print('resizing odlc')
         odlc_arr = cv.resize(odlc_arr, tuple([int(odlc_arr.shape[0]/scale), int(odlc_arr.shape[1]/scale)]))
-        odlc_arr = rotate_bound(odlc_arr, random.randrange(-180,180))
+        for key,val in curr_odlc_text_loc.items():
+            curr_odlc_text_loc[key] = int(val[0]/scale), int(val[1]/scale)
 
-        if DEBUG:
-            cv.imshow(odlc['file_name'], odlc_arr)
-            cv.waitKey(0)
-            cv.destroyAllWindows()
+        print('rotating odlc')
+        odlc_arr, rot_mat, res_size = rotate_bound(odlc_arr, random.randrange(-180,180))
+
+
+        for key,val in curr_odlc_text_loc.items():
+            curr_odlc_text_loc[key] = (
+                    int(rot_mat[0,0]*val[0] + rot_mat[0,1]*val[1] + rot_mat[0,2]),
+                    int(rot_mat[1,0]*val[0] + rot_mat[1,1]*val[1] + rot_mat[1,2])
+                    )
 
         top_left = random.randint(0, img_arr.shape[0]-odlc_arr.shape[0]), random.randint(0, img_arr.shape[1]-odlc_arr.shape[1])
+
+        for key,val in curr_odlc_text_loc.items(): 
+            curr_odlc_text_loc[key] = int(val[0] + top_left[0]), int(val[1] + top_left[1])
         
         if DEBUG:
             print((img_arr.shape[0]-odlc_arr.shape[0], img_arr.shape[1]-odlc_arr.shape[1]))
@@ -113,6 +135,7 @@ for img in empty_imgs:
             print(odlc_arr.shape)
             print(top_left)
 
+        print('putting odlc into image')
         for col in range(odlc_arr.shape[1]): 
             for row in range(odlc_arr.shape[0]):
                 if np.array(odlc_arr[row,col]).sum() < np.array((240,240,240)).sum():
@@ -129,6 +152,18 @@ for img in empty_imgs:
                         print(img_arr.shape)
                         print(insert_point[0], img_arr[insert_point[0]].shape)
                         raise err
-    cv.imshow('image '+img, img_arr)
-    cv.waitKey(0)
-    cv.destroyAllWindows()
+
+        img_name = img.split('.')[0]
+        print('inserting into', img_name+'.txt')
+        with open(os.path.join(cwd, out_dir, img_name + '.txt'), 'a') as img_file:
+            ins_string = str()
+            for key,val in curr_odlc_text_loc.items():
+                ins_string += str(val[0]) + ',' + str(val[1]) + ','
+            ins_string += odlc['alpha_num'] + '\n'
+            print(img_name+'.txt', ins_string, end='')
+            img_file.write(ins_string)
+    cv.imwrite(os.path.join(cwd, out_dir, img,), img_arr)
+    if DEBUG:
+        cv.imshow('image '+img, img_arr)
+        cv.waitKey(0)
+        cv.destroyAllWindows()
